Extract smith fuzzer into a module
I want to add a separate binary to work with the smith fuzzer, to print
the output program, to use that to seed the other fuzzer, and also to
record coverage of the fuzz corpus.
